converted ops to module and standardized use of async/await
