+++ src/ops/hailHydra.js
@@ -1,0 +1,12 @@
+const fs = require("fs");
+
+module.exports = (msg) => {
+  const memeDir = "./src/assets";
+  fs.readdir(memeDir, (err, hydraMemes) => {
+    if (err) {
+      throw err;
+    }
+    const memeIndex = Math.floor(Math.random() * hydraMemes.length);
+    msg.channel.send("", { files: [`./src/assets/${hydraMemes[memeIndex]}`] });
+  });
+};
