+++ src/ops/index.js
@@ -1,0 +1,15 @@
+const foaas = require("./foaas");
+const github = require("./github");
+const devToArticles = require("./devToArticles");
+const jinx = require("./jinx");
+const mdn = require("./mdn");
+const hailHydra = require("./hailHydra");
+
+module.exports = {
+  foaas,
+  github,
+  devToArticles,
+  jinx,
+  mdn,
+  hailHydra,
+};
