+++ src/ops/jinx.js
@@ -1,0 +1,3 @@
+module.exports = (channel) => {
+  channel.send("1\r2\r3\r4\r5\r6\r7\r8\r9\r10\rYou owe me a coke!!");
+};
